requirements.txt
langchain
langchain-community
langchain-ollama
langchain-text-splitters
chromadb
pypdf
python-dotenv
streamlit
